+++ src/termclip.c
@@ -93,6 +93,8 @@ clip_addchar(clip_workbuf * b, wchar chr, cattr * ca, bool tabs, ulong sizehint)
 static clip_workbuf *
 get_selection(bool attrs, pos start, pos end, bool rect, bool allinline, bool with_tabs)
 {
+  //printf("get_selection attrs %d all %d tabs %d\n", attrs, allinline, with_tabs);
+
   clip_workbuf *buf = newn(clip_workbuf, 1);
   *buf = (clip_workbuf){.with_attrs = attrs,
                         .capacity = 0, .len = 0, .text = 0, .cattrs = 0};
@@ -113,8 +115,21 @@ get_selection(bool attrs, pos start, pos end, bool rect, bool allinline, bool wi
     bool nl = false;
     termline *line = fetch_line(start.y);
 
-    if (start.y == cterm->curs.y) {
-      line->chars[cterm->curs.x].attr.attr |= TATTR_ACTCURS;
+    if (allinline) {
+      // this tweak (commit 975403 "export HTML: consider cursor", 2.9.1)
+      // causes cursor artefacts in connection with ClicksPlaceCursor=yes
+      // now guarded to cases of HTML copy/export
+      if (start.y == cterm->curs.y) {
+        line->chars[cterm->curs.x].attr.attr |= TATTR_ACTCURS;
+      }
+    }
+    if (allinline) {
+      // this tweak (commit 975403 "export HTML: consider cursor", 2.9.1)
+      // causes cursor artefacts in connection with ClicksPlaceCursor=yes
+      // now guarded to cases of HTML copy/export
+      if (start.y == cterm->curs.y) {
+        line->chars[cterm->curs.x].attr.attr |= TATTR_ACTCURS;
+      }
     }
 
     pos nlpos;
@@ -428,7 +443,7 @@ term_get_text(bool all, bool screen, bool command)
     else {
       termline * line = fetch_line(y);
       if (line->lattr & LATTR_MARKED) {
-        //printf("incr %d (sbtop %d/%d rows %d)\n", y, sbtop, term.sblines, term.rows);
+        //printf("incr %d (sbtop %d/%d rows %d)\n", y, sbtop, cterm->sblines, cterm->rows);
         if (y > sbtop) {
           y--;
           end = (pos){y, cterm->cols, 0, 0, false};